Added skeleton for the paper and Issue #19. Please claim a secton you would like to write. We can take many parts from the cikm paper. Also here is a note from siam about what they are looking for --> In SDM13, in addition to traditional work related to the design, analysis, and implementation of data mining algorithms and systems we also strongly encourage submissions of an applied nature. To facilitate the fair and effective reviewing of such application-oriented submissions, we encourage authors to structure such papers into the following sections.
Background and Motivation (Specific to the application domain)
Methods and Technical Solutions(Draw connections between the specific application to existing studies in the literature, clarify the constraints imposed by the application domain and distinguish your solutions from pre-existing ones in this context).
Empirical Evaluation (Self explanatory).
Significance and Impact (Provide concrete evidence of the potential significance and impact of your work in the given application domain).
The best papers in SDM13 will be invited to be expanded and submitted to a special issue in the Journal of Statistical Analysis and Data Mining.